+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ### Added:
  - Auth *register* and *registerBusiness* methods accept Regstration form object.
  - User *profile* method accepts Profile form object.
+ - Pinners *followers* method returns current user's followers when used without arguments.
  - Boards *forMe* method to get boards of current logged-in user.
 
 ### Fixed: